pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyl10nc"
version = "1.0.0"
description = "Python localization converter"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Shasnow", email = "yukikage@qq.com"}
]
keywords = ["localization", "i18n", "toml", "code-generation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/Shasnow/pyl10nc"
Repository = "https://github.com/Shasnow/pyl10nc"
Issues = "https://github.com/Shasnow/pyl10nc/issues"

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]

[project.scripts]
pyl10nc = "pyl10nc.__main__:main"
